Add token_file configuration option
This option provides an alternative way to set the access token. Instead
of providing the value in the configuration file, users can now use a
secure path to provide the secret in and let listenbrainz-mpd load it
from there.

This is especially useful for Nix users, as this allows storing the
secret outside of the world readable Nix store.

Signed-off-by: Sefa Eyeoglu <contact@scrumplex.net>